fix for unit test not working the same on UNIX as windows.

+++ core/src/test/java/org/bitcoinj/core/LevelDbFullPrunedBlockChainTest.java
@@ -25,8 +25,10 @@ public class LevelDbFullPrunedBlockChainTest extends AbstractFullPrunedBlockChai
 
     private void deleteFiles() {
         File f=new File("test-leveldb");
-        for (File c : f.listFiles())
-        	c.delete();
+        if(f!=null){
+	        for (File c : f.listFiles())
+	        	c.delete();
+        }
     }
 
     @Override
